[Calcolo #1] maxDistanceFromStart usando la formula dell'emisenoverso: sposta file
